+++ unity/Holo/Assets/Scripts/STLImportUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9bf391cf847d2334b928b29e2c625999
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
