+++ unity/Holo/Assets/Scripts/STLSeriesImporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 50d557e3ce515cb40987b3b2e45d9249
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
